9627 numeric range field
